refactor(up): extract image-acquisition step into up/images.ts (LOC budget)
The --release addition pushed up.ts over the 1100-line phase0 LOC cap.
Decompose the whole 'Step 6: Get images into ACR' block (release import /
--build / source-ACR import) into a focused up/images.ts::acquireImages()
helper. up.ts 1171 -> 924; images.ts 293 (new-file cap 800). No behaviour
change; 802 tests pass.
